test: move POST /projects test to separate file
